+++ odoo/songs/install/accounting.py
@@ -58,6 +58,14 @@ def default_values(ctx):
         'value_unpickle': 'random',
     })
 
+    account_612031 = ctx.env.ref('scenario.account_612031')
+    tax_xml_id = '__setup__.account_tax_none_main_company_' \
+                 'Frais_de_voiture___TVA_50%_Non_Deductible'
+    create_or_update(ctx, 'account.tax', tax_xml_id, {
+        'account_id': account_612031.id,
+        'refund_account_id': account_612031.id
+    })
+
 
 @anthem.log
 def company_settings(ctx):
@@ -128,6 +136,20 @@ def import_account_journal(ctx):
 
 
 @anthem.log
+def import_account_analytic_tag(ctx):
+    """ Importing account analytic tags """
+    content = resource_stream(req, 'data/install/account.analytic.tag.csv')
+    load_csv_stream(ctx, 'account.analytic.tag', content, delimiter=',')
+
+
+@anthem.log
+def import_account_analytic_account(ctx):
+    """ Importing account analytic account """
+    content = resource_stream(req, 'data/install/account.analytic.account.csv')
+    load_csv_stream(ctx, 'account.analytic.account', content, delimiter=',')
+
+
+@anthem.log
 def company_currency(ctx):
     """ Setting company's currency """
     company = ctx.env.ref('base.main_company')
@@ -150,11 +172,6 @@ def create_financial_journals(ctx):
         {'xmlid': 'scenario.expense_journal',
          'name': 'Expenses',
          'code': 'EXP',
-         'type': 'purchase',
-         },
-        {'xmlid': 'scenario.wage_journal',
-         'name': 'Wage',
-         'code': 'WAG',
          'type': 'purchase',
          },
     ]
@@ -278,15 +295,17 @@ def setup_sequences(ctx):
 def main(ctx):
     """ Configuring accounting """
     import_banks(ctx)
+    add_xmlid_account(ctx)
+    adapt_chart_of_account(ctx)
     import_account_journal(ctx)
+    import_account_analytic_tag(ctx)
+    import_account_analytic_account(ctx)
     company_settings(ctx)
     company_currency(ctx)
     activate_multicurrency(ctx)
     create_financial_journals(ctx)
-    add_xmlid_account(ctx)
     add_xmlid_tax(ctx)
     add_xmlid_fiscal_position(ctx)
-    adapt_chart_of_account(ctx)
     settings(ctx)
     default_values(ctx)
     setup_sequences(ctx)